update: close project action
